+++ src/WriteBuffer.d.ts
@@ -1,4 +1,0 @@
-import {Writable} from 'stream';
-export default class WriteBuffer extends Writable {
-  toString(encoding: string): string;
-}