+++ test/WriteBuffer.test.js
@@ -1,5 +1,5 @@
 const test = require('ava');
-const WriteBuffer = require('../src/WriteBuffer');
+const { WriteBuffer } = require('../src/WriteBuffer');
 
 const { Writable } = require('stream');
 
@@ -15,4 +15,14 @@ test('A buffer can encode the collected data as a string', async (test) => {
   buffer.write('world');
 
   test.is(buffer.toString('utf8'), 'hello world');
+});
+
+test('the buffer can be reset', async (test) => {
+  const buffer = new WriteBuffer();
+  buffer.write('aGVsbG8=', 'base64');
+  buffer.write('20', 'hex');
+  buffer.reset();
+  buffer.write('world');
+
+  test.is(buffer.toString('utf8'), 'world');
 });